Simplify detecting advanced build
In CLJ, CSS class and var names are always prepended and appended with
magic words. in CLJS, only during an advanced build so that those names
are easy to read when developping.

Previously, this was handled by setting a Java property. Now, the user
does not have to prepare anything.