Stringify resources like class names depending on context
It was cumbersome to have only one string interpretation. For instance,
when used with the DOM, class names should be mere name whereas they
should be dotted when expressing selectors in CSS rules. Class names and
IDs (added in this commit) are now objects which output mere names when
using `str` and selector names when used with `templ` (formerly `sel`).