Greatly improve hot reloading and add rule defs
A new macro for defining CSS rules serve to ensure that everything
disappear when compiling an advanced CLJS build and remove the need for
gathering rules. Any defined rule is automatically live reloaded.

Furthemore, the whole hot reloading experience is now close to perfect.
When a rule is defined, an actual file is generated. This file is easily
inspectable in the browser in order to see the output CSS. Such files
are neatly organized by mimicking definitions. Each filename is the name
of the rule symbol place in a folder corresponding to the namespace were
the definition occurs.

However, this commit currently breaks generating a main CSS file from
Clojure.