Prevent CLJS API from requiring Garden
All utilities leveraging Garden are now strictly in Clojure JVM. Macros
exist for development for things like testing `interpolate` or `templ`
at the REPL regardless of the platform.

The rationale is that there is no need for Garden in CLJS. Requiring it,
on the other hand, prevents quite a bit of dead code elimination.